requirements.txt
aioconsole~=0.8.1
#discord.py
pytest~=8.3.4
pytest-asyncio~=0.25.0
python-binance~=1.0.25
python-telegram-bot~=21.9 
PyYAML~=6.0.2
quart~=0.19.9
telethon~=1.38.1
